requirements-2025.txt
# Sophia Intel AI - 2025 Ultra-Performance Stack

# Core Frameworks
agno==1.8.1  # Ultra-fast agent framework with <2μs instantiation (pinned to latest stable)
aiobreaker>=1.0.0  # Async circuit breakers for resilience
aiocache>=0.12.2  # Async caching with Redis support
aiofiles>=23.2.1  # Async file operations
authlib>=1.3.0  # OAuth 2.0 with Resource Indicators
httpx[http2]>=0.27.0  # Async HTTP client with HTTP/2 support
locust>=2.17.0  # Load testing for 10,000+ agents

# MCP & Security
mcp>=0.1.0  # Model Context Protocol

# Observability
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
orjson>=3.9.10  # Fastest JSON serialization
passlib[bcrypt]>=1.7.4  # Password hashing
portkey-ai>=1.3.2  # Unified LLM routing
prometheus-client>=0.19.0
pydantic>=2.5.0  # Data validation with V2 performance
PyJWT>=2.8.0  # JWT handling without ecdsa dependency (cryptography-backed)

# Testing & Benchmarking
pytest-asyncio>=0.21.1
pytest-benchmark>=4.0.0
structlog>=24.1.0  # Structured logging

# Async & Performance
uvloop>=0.19.0  # Ultra-fast async event loop
weaviate-client>=4.16.4  # Advanced vector database with auto-vectorization
